+++ src/index.tsx
@@ -1,9 +1,7 @@
 import { createWeb3ReactRoot, Web3ReactProvider } from '@web3-react/core'
 import 'inter-ui'
 import React, { StrictMode } from 'react'
-import { isMobile } from 'react-device-detect'
 import ReactDOM from 'react-dom'
-import ReactGA from 'react-ga'
 import { Provider } from 'react-redux'
 import { HashRouter } from 'react-router-dom'
 import Blocklist from './components/Blocklist'
@@ -20,6 +18,7 @@ import TransactionUpdater from './state/transactions/updater'
 import UserUpdater from './state/user/updater'
 import ThemeProvider, { FixedGlobalStyle, ThemedGlobalStyle } from './theme'
 import getLibrary from './utils/getLibrary'
+import { initAnalytics } from './analytics/index'
 
 const Web3ProviderNetwork = createWeb3ReactRoot(NetworkContextName)
 
@@ -65,49 +64,3 @@ ReactDOM.render(
 )
 
 serviceWorkerRegistration.unregister()
-
-function initAnalytics(): void {
-  initFathom()
-  initGA()
-}
-
-function initFathom(): void {
-  const FATHOM_ID: string | undefined = process.env.REACT_APP_FATHOM_ID
-  if (typeof FATHOM_ID === 'string') {
-    const head = document.getElementsByTagName('head')[0]
-    const fathomScript = document.createElement('script')
-    fathomScript.type = 'text/javascript'
-    fathomScript.src = 'https://cdn.usefathom.com/script.js'
-    fathomScript.dataset.site = FATHOM_ID
-    head.appendChild(fathomScript)
-  }
-}
-
-function initGA(): void {
-  const GOOGLE_ANALYTICS_ID: string | undefined = process.env.REACT_APP_GOOGLE_ANALYTICS_ID
-  if (typeof GOOGLE_ANALYTICS_ID === 'string') {
-    ReactGA.initialize(GOOGLE_ANALYTICS_ID, {
-      gaOptions: {
-        storage: 'none',
-        storeGac: false
-      }
-    })
-    ReactGA.set({
-      anonymizeIp: true,
-      customBrowserType: !isMobile
-        ? 'desktop'
-        : 'web3' in window || 'ethereum' in window
-        ? 'mobileWeb3'
-        : 'mobileRegular'
-    })
-  } else {
-    ReactGA.initialize('test', { testMode: true, debug: true })
-  }
-
-  window.addEventListener('error', error => {
-    ReactGA.exception({
-      description: `${error.message} @ ${error.filename}:${error.lineno}:${error.colno}`,
-      fatal: true
-    })
-  })
-}
